package/platform/src/Oro/Bundle/EmailBundle/Resources/config/process.yml
package/platform/src/Oro/Bundle/ImapBundle/Command/Cron/EmailSyncCommand.php
package/platform/src/Oro/Bundle/ImapBundle/Tests/Functional/Cron/EmailSyncCommandTest.php

+++ src/Oro/Bundle/WindowsBundle/Resources/public/js/widget/dialog-widget.js
@@ -26,7 +26,8 @@ define(function(require) {
             type: 'dialog',
             dialogOptions: null,
             stateEnabled: true,
-            incrementalPosition: true
+            incrementalPosition: true,
+            preventModelRemoval: false
         }),
 
         // Windows manager global variables
@@ -38,6 +39,7 @@ define(function(require) {
         defaultPos: 'center center',
         openedWindows: 0,
         contentTop: null,
+        keepAliveOnClose: false,
         /**
          * Flag if the widget is embedded to the page
          * (dialog has own life cycle)
@@ -128,7 +130,9 @@ define(function(require) {
             if (_.isFunction(onClose)) {
                 onClose();
             }
-            this.dispose();
+            if (!this.keepAliveOnClose) {
+                this.dispose();
+            }
         },
 
         /**
@@ -139,7 +143,7 @@ define(function(require) {
                 return;
             }
             dialogManager.remove(this);
-            if (this.model) {
+            if (this.model && !this.options.preventModelRemoval) {
                 this.model.destroy({
                     error: _.bind(function(model, xhr) {
                         // Suppress error if it's 404 response and not debug mode
@@ -153,6 +157,8 @@ define(function(require) {
 
             // need to remove components in widget before DOM will be deleted
             this.disposePageComponents();
+            _.invoke(this.subviews, 'dispose');
+
             if (this.widget) {
                 this.widget.remove();
                 delete this.widget;
@@ -238,7 +244,7 @@ define(function(require) {
          * Removes dialog widget
          */
         remove: function() {
-            if (this.widget) {
+            if (this.widget && this.widget.dialog('isOpen')) {
                 // There's widget, close it before remove.
                 // Close handler will invoke dispose method,
                 // where remove method will be called again
@@ -273,6 +279,7 @@ define(function(require) {
 
         _clearActionsContainer: function() {
             this.widget.dialog('actionsContainer').empty();
+            this.actionsEl = null;
         },
 
         _renderActions: function() {
@@ -298,6 +305,9 @@ define(function(require) {
                 this.widget.html(this.$el).dialog(dialogOptions);
                 this.getLayoutElement().attr('data-layout', 'separate');
             } else {
+                if (this.widget.dialog('instance') !== void 0 && !this.widget.dialog('isOpen')) {
+                    this.widget.dialog('open');
+                }
                 this.widget.html(this.$el);
             }
             this.loadingElement = this.$el.closest('.ui-dialog');
@@ -312,6 +322,14 @@ define(function(require) {
                 this._fixDialogMinHeight(false);
                 this.widget.trigger('resize');
             }, this));
+        },
+
+        hide: function() {
+            // keepAliveOnClose property is used to avoid disposing the widget on dialog close to be able open it again
+            var keepAliveOnClose = this.keepAliveOnClose;
+            this.keepAliveOnClose = true;
+            this.widget.dialog('close');
+            this.keepAliveOnClose = keepAliveOnClose;
         },
 
         _renderHandler: function() {
